+++ R/class_timestamp.R
@@ -30,7 +30,7 @@ timestamp_class <- R6::R6Class(
         time_stamp = TRUE
       )
     },
-    report_cancelled = function(target = NULL, progress = NULL) {
+    report_canceled = function(target = NULL, progress = NULL) {
       cli_cancel(
         target_get_name(target),
         target_get_type_cli(target),