+++ tests/testthat/test-class_stem.R
@@ -276,7 +276,7 @@ tar_test("branches can use old buds if continuing on error", {
   expect_equal(tar_read(y), seq_len(3))
 })
 
-tar_test("branches can use old buds if stem is cancelled", {
+tar_test("branches can use old buds if stem is canceled", {
   tar_script({
     list(
       tar_target(x, seq_len(3)),
@@ -298,7 +298,7 @@ tar_test("branches can use old buds if stem is cancelled", {
   expect_equal(tar_read(y), seq_len(3))
 })
 
-tar_test("branches can use old buds if stem is cancelled (worker storage)", {
+tar_test("branches can use old buds if stem is canceled (worker storage)", {
   tar_script({
     list(
       tar_target(x, seq_len(3), storage = "worker"),